Move FilterOption to Contracts namespace level
